Moved a util function to searches.py since python 2 doesn't like circular imports
