+++ cs4660/graph/utils.py
@@ -105,25 +105,3 @@ def convert_edge_to_grid_actions(edges):
             path_string = path_string + direction_dict[(tile1.x-tile2.x,tile1.y-tile2.y)]
 
     return path_string
-
-def get_path_to_destination_node(graph, node_to_parent_dict, dest_node):
-    """
-    Returns the path to the from the root node to the dest_node using 
-    the node->parent dictionary provided. The dictionary is created 
-    as a by-product of any graph search algorithm
-    """
-    path = []
-
-    if dest_node in node_to_parent_dict:
-        # Start with the destination node to compute the path upward
-        parent_node = node_to_parent_dict[dest_node]
-        while parent_node is not None:
-            # Add the edge to the top of list since we move bottom up
-            path.insert(0, g.Edge(parent_node, dest_node, graph.distance(parent_node, dest_node)))
-            dest_node = parent_node
-            if dest_node in node_to_parent_dict:
-                parent_node = node_to_parent_dict[dest_node]
-            else:
-                parent_node = None
-
-    return path
